Remove the world Minecraft from the fist page.
One problem with the voxel type of games is that most people will think
"a Minecraft clone", that's fine, we are inspired by modded Minecraft,
Infiniminer, Minetest and similar type of block building sandbox type of
games. Still, maybe not throw the world "minecraft" in the face at out
new visitors :)

I choose to use "block building" because that's what the current game
mostly is about. When we have mobs and stuff we can update the title.